Backend/requirements.txt
# Web framework
fastapi==0.135.1
uvicorn==0.42.0
apscheduler==3.10.4

# Database
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4

# Config & validation
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1

# HTTP clients (external API calls)
requests==2.32.5
httpx==0.28.1
feedparser==6.0.12

# Machine learning
scikit-learn==1.8.0
xgboost==3.1.3
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1